+++ pytests/test_common.py
@@ -7,20 +7,20 @@
 def test_package_import() -> None:
     """Test that the package can be imported."""
     # pylint: disable=import-outside-toplevel
-    from frequenz.api import common
+    from frequenz.api.common import v1
 
-    assert common is not None
+    assert v1 is not None
 
 
 def test_module_import_components() -> None:
     """Test that the modules can be imported."""
     # pylint: disable=import-outside-toplevel
-    from frequenz.api.common import components_pb2
+    from frequenz.api.common.v1 import components_pb2
 
     assert components_pb2 is not None
 
     # pylint: disable=import-outside-toplevel
-    from frequenz.api.common import components_pb2_grpc
+    from frequenz.api.common.v1 import components_pb2_grpc
 
     assert components_pb2_grpc is not None
 
@@ -28,12 +28,12 @@ def test_module_import_components() -> None:
 def test_module_import_metrics() -> None:
     """Test that the modules can be imported."""
     # pylint: disable=import-outside-toplevel
-    from frequenz.api.common import metrics_pb2
+    from frequenz.api.common.v1 import metrics_pb2
 
     assert metrics_pb2 is not None
 
     # pylint: disable=import-outside-toplevel
-    from frequenz.api.common import metrics_pb2_grpc
+    from frequenz.api.common.v1 import metrics_pb2_grpc
 
     assert metrics_pb2_grpc is not None
 
@@ -41,11 +41,11 @@ def test_module_import_metrics() -> None:
 def test_module_import_metrics_electrical() -> None:
     """Test that the modules can be imported."""
     # pylint: disable=import-outside-toplevel
-    from frequenz.api.common.metrics import electrical_pb2
+    from frequenz.api.common.v1.metrics import electrical_pb2
 
     assert electrical_pb2 is not None
 
     # pylint: disable=import-outside-toplevel
-    from frequenz.api.common.metrics import electrical_pb2_grpc
+    from frequenz.api.common.v1.metrics import electrical_pb2_grpc
 
     assert electrical_pb2_grpc is not None